win: Fix always-rebuild of crashpad_snapshot_test_image_reader_module.dll
Ninja assumes all DLLs will have an import library generated (caused
when there are any exports), but because this DLL is so simple, it does
not. This makes ninja think that the target is always dirty and so it
rebuilds it on every build. Fix this by telling ninja not to expect an
import library.

R=mark@chromium.org
BUG=crashpad:1

Review URL: https://codereview.chromium.org/1346253003 .