Ensure _NT_SYMBOL_PATH is set for bot runs in cdb test
Getting closer... Some tests passed on the last run, but the ones that
rely on having ntdll symbols fail on the bot. With `_NT_SYMBOL_PATH`
set, cdb will be able to download the PDBs so will be able to dump
data for `ntdll!_PEB`, etc.

R=mark@chromium.org
BUG=crashpad:46

Review URL: https://codereview.chromium.org/1402643002 .