Update patch set 2
Patch Set 2:

This is really becoming less and less Windows-specific. These concepts apply equally to any platform. Are you sure that you want to continue down the path of doing this kind of thing as Windows-only? It seems like that’s just going to make for more work when we eventually convert it to not be Windows-only.

Also, what about a provision for dialing the captured extra data up or down, as was mentioned on the mailing list thread?

Patch-set: 2