Update patch set 2
Patch Set 2:

Scott Graham wrote:
> Partly due to panic-driven-development, and partly because I didn't
> have a handy way to test on non-Windows.

I’m just afraid that as we dig ourselves deeper into a hole, it’ll be harder for us to climb back out. (I’ve also got a fear is that it’ll be harder for me to climb out alone, hopefully that’s unfounded.)

> Yeah, I was looking at that. I guess it's a per-client thing, so on
> Windows I'd pass it through in RegistrationRequest, and then pass the
> level through to ExceptionHandlerServerException. Sound OK?

Hang it off of the CrashpadInfo struct/section? That’ll plug in more cleanly to the design on some platforms where registration isn’t as explicit or wordy, and it’ll make things easier on others by not requiring that this knob be piped all the way through to the necessarily-distinct per-platform registration code.

(We’ve also learned that Gerrit, at least as we’ve gotten it configured right now, doesn’t pick up on comments sent by e-mail.)

Patch-set: 2