Update patch set 7
Patch Set 7:

(9 comments)

Thanks!

Siggi separately pointed out that we have a race in both this IPC and the registration: If the client crashes immediately after sending the message but before we get to the OpenProcess() call, then the PID could have been replaced by another process, and we could open the wrong process. So we should amend the IPC message to include the process start time in addition to the PID, and have the handler verify that the start time is the expected one.

Patch-set: 7
Label: Code-Review=0