win: Add dump-and-crash for arbitrary target
This allows a client to crash another program, given the handle. The
message takes a process HANDLE and duplicates it into the handler with
SAME_ACCESS. This is intended to prevent escalation in that the handler
is only doing things that client could already do itself.

Mostly a bunch of plumbing here because the exception record is being
created rather than being read from the process that's being crashed.

Change-Id: I7f30bf2ffe85516e4052617e7ddbe1dddfe3fef9

BUG=

Change-Id: I91589ab8b0e55fe7adc5141e1fd236615789806b