Update patch set 14
Patch Set 14:

> Patch Set 14:
> 
> You know, rather than having the user invent a fake thread to blame at all, we could just write an exception-less minidump. Is there a reason to try to blame a single thread?

Yeah, we can certainly do that too (in fact, I think passing a thread id of 0 will do that now). I believe the intention was that the hang-watcher knows enough (maybe?) to blame a particular thread, and so in the crash UI, it'd be nice to have that one show up for signatures, etc.

Patch-set: 14
Label: Code-Review=0