Update patch set 14
Patch Set 14:

> You know, rather than having the user invent a fake thread to blame
 > at all, we could just write an exception-less minidump. Is there a
 > reason to try to blame a single thread?

Ah, it's a good point not every user might have a culprit thread in mind.

In the case of a failed browser rendez-vous, we blame the main thread, or whatever thread it's stuck waiting on. This simplifies analysis greatly because crashes are bucketed on the crashing stack.

Patch-set: 14
Reviewer: Pierre-Antoine Manzagol <1141194@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=0