Update patch set 1
Patch Set 1:

(There's no way to edit the commit message at the moment: https://crbug.com/603207. This is what I would have added.)

This is in order to be able to move snapshot/api/module_annotations_win
to client (which is a more natural place for it) and for a new client
function, DumpAndCrashTargetProcess(), so that the client code can
retrieve the location of CrashpadInfo for another process to stash data
there.

Patch-set: 1
Label: Code-Review=0