win: Support dumping another process by causing it to crash
Stashes a thread id (and exception code) into the target process's
CrashpadInfo in the main module and the causes the target to crash. On
the handler side, when the thread id is set, fabricates an exception
record with that thread id and exception code.

One tricky bit is that the more obvious DebugBreakProcess() doesn't do
anything unless a debugger is attached. TODO