win: Support dumping another process by causing it to crash
As a first step, stashes the blamee thread id into the target process's
CrashpadInfo (in the main module) but does not yet use it.

One tricky bit is that the more obvious DebugBreakProcess() does not
seem to do anything (at least on Win10). I suspect it's probably
something to do with the bonus thread creation flags that make DllMain
and TLS initialization not to run (so the UnhandledExceptionFilter isn't
set up?) but I'm not certain. Standalone test at
https://gist.github.com/sgraham/79f6b99abaf46591716d0c59cec30ef3. So,
instead we CreateRemoteThread() with a thread entry point of 0 to cause
a crash.

BUG=crashpad:103

Change-Id: I797007bd2b1e3416afe3f37a6566c0cdb259b106
Reviewed-on: https://chromium-review.googlesource.com/339263