Update patch set 2
Patch Set 2:

(1 comment)

Is this actually useful given the finding in https://crashpad.chromium.org/bug/103#c5?

Is the goal here to use this code from client to find an address in the victim’s memory map to *write* to? This is a little weird, it adds a requirement that the victim be a Crashpad client, which we may have wanted to enforce anyway, but it does it in a weird and ambiguous find-the-module sort of way.

It’s also kind of weird that I think the equivalent Mac code used to live in util, but we had a reason to move it into snapshot. Now the Windows code started out in snapshot and needs to move to util. I’m starting to take this as a sign that our layering isn’t working for us.

Patch-set: 2