Update patch set 4
Patch Set 4:

I see where you’re heading. Another concern: would GetProcAddress() take the loader lock, or would that just be LoadLibrary()/GetModuleHandle()?

Now that we’re talking about this, does NtCreateThreadEx() create the remote thread in the native bitness of the caller or does it set up a 32-bit environment in a WoW64 process? I assume that the unhandled exception filter runs on the created thread, a WoW64 process would have a 32-bit UEF, and I hope that this all works out cleanly.

Patch-set: 4