Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> I see where you’re heading. Another concern: would GetProcAddress() take the loader lock, or would that just be LoadLibrary()/GetModuleHandle()?

Thinking about it some more, we should be able to resolve this without running code to do so in the target. We can get the base address of kernel32 in the other bitness by reading the \KnownDlls32\kernel32.dll section (or probably other ways) and then we could walk through the export table to get the function address. In fact it looks like one "MrNukealizer" has already done so http://www.codeproject.com/Tips/139349/Getting-the-address-of-a-function-in-a- DLL-loaded . i.e. just reimplementing GetProcAddress() but we could be confident it wasn't taking the loader lock in-process, anyway.

But having said that, I'll go with just implementing the same-bitness version since that should be relatively straightforward, and since Chrome will be the only immediate consumer, and it doesn't require cross-bitness target dumping at the moment.
 
> Now that we’re talking about this, does NtCreateThreadEx() create the remote thread in the native bitness of the caller or does it set up a 32-bit environment in a WoW64 process? I assume that the unhandled exception filter runs on the created thread, a WoW64 process would have a 32-bit UEF, and I hope that this all works out cleanly.

Not sure, I'd go with hoping it all just works out, but I would have to test that for cross-bitness.


So, in summary, I think I will switch to an injected RaiseException() call, and a hopefully a bit less code in the handler's InitializeException() path that's special to this case.

I'm not sure if we should still move the pe_image_reader, etc. from snapshot to util. We won't need it here now for reading CrashpadInfo. But the code that's in snapshot/api feels like it's in the wrong place and belongs in client, but can't be there now because of using those classes to read the annotations.

Patch-set: 4
Label: Code-Review=0