Update patch set 9
Patch Set 9:

Added x64 code to call to RaiseException() in the latest patch set. Unfortunately seems to skip SetUnhandledExceptionFilter() and go straight to the "program has crashed" system dialog. Argh!

Going to do a bit more testing to make sure it's not because the injection code is somehow messing up the exception handler setup, but I have the feeling that might be just how RaiseException() works on x64.

So might need to rework yet-again using ud2.

Patch-set: 9
Label: Code-Review=0