Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> P.S. I would normally suggest uint64_t for sizey things of generic bitness, but given that the goal here is to restrict what goes into a minidump file, it seems like 4GB is totally acceptable as an overall cap.

Agreed. We should support "full dump" if we ever want > 4G.
 
> P.P.S. It’d be nice if limit = some_constant (0 or max) was interpreted as “unlimited.”

That was part of the reason I used int originally, I was going to have -1 mean unlimited. But I can't see any problem in practice with specifying numeric_limits<uint32>::max() and having that be a real limit, but not strictly infinite.

Patch-set: 6
Label: Code-Review=0