Update patch set 1
Patch Set 1:

(I also realized while writing this that using ScopedSuspendProcess isn't quite right. In the destructor for that we're going to try to resume a process that we may have already terminated elsewhere. It's probably not a huge deal though as the only negative effect would be a spurious log message.)

Patch-set: 1