Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (1 comment)
> 
> I was hoping that we could do this in the ProcessReader so that all of the logic to mess with suspend_count was consolidated. We already mess with it in FillThreadContextAndSuspendCount<>(). Is that too much plumbing?

It's more that it's too "late". In order to InitializeException() we need a set of thread contexts so we've already finished having ProcessReader get all its data before we know that we need to modify the suspend count when we process the exception.

But maybe we could either "peek" at that, or alternatively move the current suspend count later... I'll take a look at that.

Patch-set: 1