Update patch set 1
Patch Set 1:

Unrelatedly, this test is insanely slow in Debug ~80s vs. ~1s in Release on my machine.

It appears to be due to too many calls to GetReadableRangesOfMemoryMap() which are perhaps much slower in Debug due to the use of CheckedNumeric. (This isn't new behaviour, it just shows up here, so I'll take a look at that separately.)

Patch-set: 1
Label: Code-Review=0