Update patch set 3
Patch Set 3:

Mark, do you have any feelings on this? My main goal is to remove the limitation of having exactly/only 64 256+256 entries, and so requiring RegisterCrashKeys() in Chrome, which has gotten increasingly ugly.

I could also implement it as a simple growable array of pairs of char*, rather than the fiddling with interning, which would just push a bit more complexity to read-time. But might be simpler overall, and would probably be easier to extend later if we decide to support more complex value types.

Patch-set: 3