mac: dyld fatal errors appear as exit(1) on 10.12
exit(1) is a weird code for this, so I filed
https://openradar.appspot.com/26894758.

This doesn’t completely fix bug crashpad:121 unless both
crashpad_snapshot_test and crashpad_snapshot_test_no_op are signed with
the same Developer ID certificate. I’m hoping to get some action on
https://openradar.appspot.com/26902656, which will enable a complete fix
for this bug in unsigned developer builds. It would be unusual to have
to sign test executables.

BUG=crashpad:120,crashpad:121
TEST=crashpad_snapshot_test MachOImageAnnotationsReader.CrashDyld

Change-Id: I54fdfaa9178029b91ea3cbc12f2760dfa5124858