Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Oh yeah, uninstaller should be fixed (long due too), but that is a moderate rewrite. Making sure cashpad_handler is never explicitly killed avoids the problem entirely, thanks.
> 
> The crashes we saw are all due to FATAL log messages calling int3 when handler fails to start after being deleted and killed. 

I meant why crashpad_handler was crashing in the first place, but you said above that your uninstaller was killing it, so that answers that.

Obviously, if it was crashing on its own, I’d want to fix the crash.

> Here is the stack for a contrived scenario of manually deleted and explicitly killed handler. This will keep repeating every second. Not sure if such scenario could occur in valid circumstances, raving uninstallers notwithstanding (flickering NFS mounts coupled with handler crash?), but maybe exec failure shouldn't be fatal and exit gracefully instead. Crashing raises problem's visibility, but creates system crash reports that could eventually fill up user's home directory. Unless there is a limit of how many restarts are attempted?

The logic that we currently use is that if the first attempt to start crashpad_handler fails, then we’ll give up right then and there. This is the CommonStart() called by InitialStart(). InitialStart() does test CommonStart()’s return value. Assuming that the initial start succeeded, InitialStart() then sets up the handler restarter if desired.

The idea is that if the first handler start succeeds and the handler crashes, then we should be able to start the handler again. But we never enable the retry logic unless we were able to at least start successfully in the first place.

So with that in mind, even if you’ve got a “flickering” NFS mount, we know that it was working enough to start the handler at one point, and if it’s flickering, maybe it’ll flicker back to a usable state.

When I wrote this up in the first place, I considered exponential backoff for the restart sleep, but decided to go with the simple 1Hz rate limit instead. If a real client crashes while handler restart is pending, that client will be blocked solidly until a viable handler shows up. And if the real client that crashes is the initial client while a restart is pending within it, then the crash will be lost entirely.

If you can easily make your uninstaller not kill crashpad_handler, that’s what works best with crashpad_handler’s own expectations. Otherwise, if the restart semantics don’t work well for your system’s design, you’re free to disable it when you call StartHandler(). That’ll give you a crashpad_handler that you can kill at will from your uninstaller or elsewhere, but obviously no protection against it crashing.

Patch-set: 1