Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I spoke to Mark about this yesterday, and he says he'd prefer to have the handler do more work to gather strings, rather than chance around with more complicated memory operations in the crashy process. With the out-of-process architecture, it shouldn't be hard to just pull a list of
> 
> struct MapEntry {
>   char* key;
>   size_t key_len;
>   char* value;
>   size_t value_len;
> };

Thanks for the review and feedback. Sorry for the delay in turnaround, was out for a while and just getting back into things now.

You're right about "reallocation time". I was thinking more about this (showerthoughts!) and it seems worth doing better, in particular on ensuring the reader always sees a consistent view of the annotations structure. I'll write up a short design to see if we can agree on the the goals and tradeoffs for this data structure.

Patch-set: 3