Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> Cute. OK, let’s go with this.
> 
> I thought that there was some link-time CRT selection thing going on with the different subsystems. Maybe there is, but it’s not really that relevant to what winds up running as long as you use the /subsystem:windows-type wWinMain entry point?

Yeah, it's a bit involved. The _linker_ uses the choice of /subsystem to determine what entry point it expects, but primarily it sets a bit in the header, which is what the kernel uses when starting the program to determine if it should get a console.

But the kernel doesn't care what the entry point is called, and sets up stdout, etc. in the same way. So we set it to /subsystem:windows so crashpad_handler.exe works as a GUI app by default, but then flipping it to console after the fact in the second copy means that it'll get a console and just work, even though it happens to be going to the function named WinMain().

(At least that's the theory... at a minimum, crashpad_handler.exe should be fine as a default for users that won't typically want to see a console, and we'll see locally over time if the .com has any weird gotchas.)

Patch-set: 3
Label: Code-Review=0