Update patch set 5
Patch Set 5:

Hi Mark,

This isn't completely finished yet, but it's a bit invasive. So, I'd appreciate if you could do a first review now to make sure you find it reasonable. I've tried a few different ways to implement this already and this is the best I came up with.

It's a bit tricky because the coalescing has to happen at the "end" because that's when all desired memory ranges are known. So minidump seems like the place to do this, but it doesn't own or know much about the snapshot types.

The precursor change is fairly mechanical, and only necessary if this one is done.

The not-finished parts are that I need to determine how coalescing interacts with the stack regions (this will just be "throw away anything that overlaps the stack" because the stack should be a complete capture anyway), and it needs more tests at the minidump level.

Patch-set: 5
Label: Code-Review=0