Update patch set 1
Patch Set 1:

The way I understand it, it's intended to send me e-mail as soon as you add me as a reviewer. This should be the case even if you don't send your own message. I'm supposed to see it as soon as I'm on the reviewer list, even if you're not ready for me yet.

Since this is all so confusing, maybe we should file another [Poly]Gerrit bug about making the intended behavior somehow more apparent in the UI. Or, of appropriate, changing the behavior to better match user expectations.

Patch-set: 1