Update patch set 7
Patch Set 7:

After a back and forth, Will would prefer a uint32_t accepting interface in Metrics. I think that's fine. Given that we're swapping back and forth between int32_t and uint32_t and exception_type_t and DWORD everywhere, it's all kind of moot, and this makes for less code in the header.

Patch-set: 7
Label: Code-Review=0